// src/question-sets/dto/create-practice-set.dto.ts
import { DifficultyLevel, PracticeSetVisibility } from '@prisma/client';

export class CreatePracticeSetDto {
  title!: string;
  description?: string;
  type!: string;
  createdBy!: string;
  visibility?: PracticeSetVisibility;
  totalMarks?: number;
  timeLimit?: number;
  estimatedDifficulty?: DifficultyLevel;
  questionIds?: string[];
  sourceFilterSnapshot?: unknown;
}